Modified num0 to be a float

+++ strcalc.R
@@ -2,9 +2,9 @@
   # e.g. '1 + 2' -> 3, or '1 - 2' -> -1
 compute = function(input_string) {
   values = unlist(strsplit(input_string, ' '))
-  num0 = as.integer(values[1])
+  num0 = as.numeric(values[1])
   operator = values[2]
-  num1 = as.float(values[3])
+  num1 = as.numeric(values[3])
   result = switch(operator,
                   '+' = num0 + num1,
                   '-' = num0 - num1,
